Perpex/NightOwl1/requirements.txt
psutil
pyyaml
rich
python-dotenv
cerberus
